+++ Engine/Source/Core/MouseCodes.h
@@ -1,9 +1,8 @@
 #pragma once
-#include <cstdint>
 
 namespace Apex {
 
-    using MouseCode = uint16_t;
+    using MouseCode = int;
 
     namespace Mouse {
